restructure component folder and fix issues
